Improve square UX and refactor konami
